* use size_t for Com_sprintf instead of int

+++ src/game/q_shared.h
@@ -342,7 +342,7 @@ char *COM_Parse(char **data_p);
 /* data is an in/out parm, returns a parsed out token */
 char *COM_EParse(char **text, char *errhead, char *errinfo);
 
-void Com_sprintf(char *dest, int size, char *fmt, ...);
+void Com_sprintf(char *dest, size_t size, char *fmt, ...);
 
 void Com_PageInMemory(byte * buffer, int size);
